pyproject.toml
[tool.poetry]
name = "home-assistant"
version = "0.1.0"
description = ""
authors = ["luikymagno <luikymagno@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
kafka-python = "^2.0.2"
aiokafka = "^0.7.2"
grpcio = "^1.50.0"
grpcio-tools = "^1.50.0"
uvicorn = "^0.19.0"
fastapi = "^0.85.1"
requests = "^2.28.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
